requirements.txt
aiodns==2.0.0
aiofiles==0.5.0
aiohttp==3.6.2
aiohttp-requests==0.1.3
aiomysql==0.0.20
aioredis==1.3.1
aiosqlite==0.13.0
alembic==1.4.2
async-timeout==3.0.1
attrs==19.3.0
cchardet==2.1.6
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cryptography==2.9.2
fastapi==0.54.1
h11==0.9.0
hiredis==1.0.1
httptools==0.1.1
idna==2.10
iso8601==0.1.12
jwt==1.0.0
lxml==4.5.1
Mako==1.1.3
MarkupSafe==1.1.1
multidict==4.7.6
passlib==1.7.2
pyasn1==0.4.8
pycares==3.1.1
pycparser==2.20
pydantic==1.5.1
PyJWT==1.6.4
PyMySQL==0.9.2
PyPika==0.37.15
python-dateutil==2.8.1
python-decouple==3.3
python-editor==1.0.4
python-multipart==0.0.5
pytz==2020.1
redis==3.5.3
rsa==4.0
six==1.15.0
SQLAlchemy==1.3.18
starlette==0.13.2
tortoise-orm==0.16.13
typing-extensions==3.7.4.2
ujson==3.0.0
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
xmltodict==0.12.0
yarl==1.4.2
ciso8601==2.1.3